iOS Text: pull enter_text and _enter_text_in into API
