+++ lib/calabash/ios/api.rb
@@ -1,8 +1,11 @@
 module Calabash
   module IOS
     module API
+      require 'calabash/gestures'
+      require 'calabash/wait'
       require 'calabash/ios/api/keyboard'
       require 'calabash/ios/api/status_bar'
+      require 'calabash/ios/api/text'
     end
   end
 end
